+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## New Features
 
-1. Added functionality to Scatter Shot PowerUp.
+1. PowerUps will now spawn randomly at random locations. They will not spawn inside walls.
 
 ## Changes
 
-1. Reduce bullet speed in order to prevent glitches.
+1. Increased time required to despawn bullets.
+2. Increased the number of bullets shot out by Scatter Shot.
+3. Increased the speed of Scatter Shot Bullets.